pyproject.toml
[project]
name = "learnforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=2.20.0",
    "langchain>=0.2.14",
    "langchain-community>=0.2.12",
    "langchain-openai>=0.1.22",
    "langchain-qdrant>=0.1.3",
    "langchain-text-splitters>=0.2.2",
    "langgraph>=0.2.20",
    "langsmith>=0.1.103",
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.10.1",
    "ragas>=0.1.20",
    "streamlit>=1.36.0",
    "tavily-python>=0.5.0",
]
